Move split_data into end_model/data.py
